Trekker Snake ut i egen katalog, egen less-fil
