+++ src/components/games/SpareslagenSpill/SpareslangenSpill.module.less
@@ -51,10 +51,6 @@
     }
 }
 
-.snake {
-    transition: background-color 0.5s;
-}
-
 .snakeLine {
     display: flex;
     margin-top: 40px;
