+++ src/components/games/SpareslagenSpill/snake/snake.module.less
@@ -1,0 +1,3 @@
+.snake {
+  transition: background-color 0.5s;
+}
